+++ package.json
@@ -89,8 +89,7 @@
       "**/@types/*/**",
       "**/grunt-*",
       "**/grunt-*/**",
-      "x-pack/typescript",
-      "kbn_tp_*/**"
+      "x-pack/typescript"
     ]
   },
   "dependencies": {
